chore(refactor) : Refactor game logic into dedicated hooks for modularity and maintainability
- Create useGameBoardManagement, useGameStatus, and useWordManagement hooks.
- Update useGameState
